+++ assets/panel/moz.build
@@ -7,6 +7,7 @@ DevToolsModules(
     'debugger.css',
     'debugger.js',
     'panel.js',
+    'parser-worker.js',
     'pretty-print-worker.js',
     'source-map-worker.js'
 )
