+++ src/utils/parser/worker.js
@@ -1,0 +1,8 @@
+const { getSymbols, getVariablesInScope } = require("./utils");
+
+const { workerHandler } = require("../worker");
+
+self.onmessage = workerHandler({
+  getSymbols,
+  getVariablesInScope
+});
