+++ src/utils/pretty-print.js
@@ -1,7 +1,7 @@
 // @flow
 
 const { getValue } = require("devtools-config");
-const { workerTask } = require("./utils");
+const { workerTask } = require("./worker");
 const { isJavaScript } = require("./source");
 const assert = require("./assert");
