+++ webpack.config.js
@@ -19,7 +19,7 @@ function buildConfig(envConfig) {
     entry: {
       debugger: getEntry("main.js"),
       "source-map-worker": getEntry("utils/source-map-worker.js"),
-      "parser-worker": getEntry("utils/parser/index.js"),
+      "parser-worker": getEntry("utils/parser/worker.js"),
       "pretty-print-worker": getEntry("utils/pretty-print-worker.js"),
       "integration-tests": getEntry("test/integration/tests.js"),
     },
